NMS-10552: Slack-compatible notification strategies expect "url" for switch name, should be "-url"
Fix it so leading dashes are expected. Include a warning message if we see a switch name without
a leading dash, so that people can be prompted to update their notification command definitions.
Since we have never shipped a config with such a command on by default, this warning should do
the job.